Fix warning (when compiling with strict warnings).
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@3505 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/uqm/netplay/packet.h
@@ -108,7 +108,7 @@ packetLength(const Packet *packet) {
 
 static inline PacketType
 packetType(const Packet *packet) {
-	return (PacketType) ntoh16(packet->header.type);
+	return (PacketType) (int) ntoh16(packet->header.type);
 }
 
 static inline bool
